backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
paho-mqtt
email-validator
psycopg2-binary
aiocoap==0.4.8